Move `Parser` package
